--- journal/src/main/java/io/camunda/zeebe/journal/file/record/CorruptedLogException.java
@@ -1,33 +1,0 @@
-/*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.zeebe.journal.file.record;
-
-import io.camunda.zeebe.util.exception.UnrecoverableException;
-
-public final class CorruptedLogException extends UnrecoverableException {
-
-  public CorruptedLogException(final String message) {
-    super(message);
-  }
-
-  public CorruptedLogException(final Throwable cause) {
-    super(cause);
-  }
-
-  public CorruptedLogException(final String message, final Throwable cause) {
-    super(message, cause);
-  }
-}
