+++ journal/src/main/java/io/camunda/zeebe/journal/record/CorruptedLogException.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Zeebe Community License 1.1. You may not use this file
+ * except in compliance with the Zeebe Community License 1.1.
+ */
+package io.camunda.zeebe.journal.record;
+
+import io.camunda.zeebe.util.exception.UnrecoverableException;
+
+public final class CorruptedLogException extends UnrecoverableException {
+
+  public CorruptedLogException(final String message) {
+    super(message);
+  }
+
+  public CorruptedLogException(final Throwable cause) {
+    super(cause);
+  }
+
+  public CorruptedLogException(final String message, final Throwable cause) {
+    super(message, cause);
+  }
+}
