+++ journal/src/main/java/io/camunda/zeebe/journal/record/InvalidRecordException.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Zeebe Community License 1.1. You may not use this file
+ * except in compliance with the Zeebe Community License 1.1.
+ */
+package io.camunda.zeebe.journal.record;
+
+public class InvalidRecordException extends RuntimeException {
+
+  public InvalidRecordException(final String message) {
+    super(message);
+  }
+}
