+++ journal/src/main/java/io/camunda/zeebe/journal/record/SBESerializer.java
@@ -1,19 +1,11 @@
 /*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Zeebe Community License 1.1. You may not use this file
+ * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.journal.file.record;
+package io.camunda.zeebe.journal.record;
 
 import io.camunda.zeebe.journal.file.MessageHeaderDecoder;
 import io.camunda.zeebe.journal.file.MessageHeaderEncoder;
@@ -117,6 +109,7 @@ public final class SBESerializer implements JournalRecordSerializer {
     return new RecordData(recordDecoder.index(), recordDecoder.asqn(), data);
   }
 
+  @Override
   public int getMetadataLength(final DirectBuffer buffer, final int offset) {
     headerDecoder.wrap(buffer, offset);
     return headerDecoder.encodedLength() + headerDecoder.blockLength();
