+++ journal/src/test/java/io/camunda/zeebe/journal/JournalTest.java
@@ -14,10 +14,10 @@ import io.camunda.zeebe.journal.JournalException.InvalidChecksum;
 import io.camunda.zeebe.journal.JournalException.InvalidIndex;
 import io.camunda.zeebe.journal.file.SegmentedJournal;
 import io.camunda.zeebe.journal.file.SegmentedJournalBuilder;
-import io.camunda.zeebe.journal.file.record.CorruptedLogException;
-import io.camunda.zeebe.journal.file.record.PersistedJournalRecord;
-import io.camunda.zeebe.journal.file.record.RecordData;
-import io.camunda.zeebe.journal.file.record.RecordMetadata;
+import io.camunda.zeebe.journal.record.CorruptedLogException;
+import io.camunda.zeebe.journal.record.PersistedJournalRecord;
+import io.camunda.zeebe.journal.record.RecordData;
+import io.camunda.zeebe.journal.record.RecordMetadata;
 import io.camunda.zeebe.journal.util.LogCorrupter;
 import io.camunda.zeebe.journal.util.TestJournalRecord;
 import io.camunda.zeebe.util.buffer.BufferUtil;
