+++ journal/src/test/java/io/camunda/zeebe/journal/util/LogCorrupter.java
@@ -10,8 +10,8 @@ package io.camunda.zeebe.journal.util;
 import io.camunda.zeebe.journal.file.FrameUtil;
 import io.camunda.zeebe.journal.file.MessageHeaderDecoder;
 import io.camunda.zeebe.journal.file.SegmentDescriptor;
-import io.camunda.zeebe.journal.file.record.JournalRecordReaderUtil;
-import io.camunda.zeebe.journal.file.record.SBESerializer;
+import io.camunda.zeebe.journal.record.JournalRecordReaderUtil;
+import io.camunda.zeebe.journal.record.SBESerializer;
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.File;
